Issue #609: Refactor ConfigStore file.
